kapitan_chat_backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf_spectacular
channels
uvicorn
websockets